pyproject.toml
[project]
name = "dsp_mapping"
version = "0.1.0"
description = ""
authors = [
    {name = "tom-billington",email = "tm.billington@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=2.3.0,<3.0.0)",
    "google-cloud-bigquery (>=3.34.0,<4.0.0)",
    "db-dtypes (>=1.4.3,<2.0.0)",
    "google-api-python-client (>=2.171.0,<3.0.0)",
    "pylint (>=3.3.7,<4.0.0)",
    "fastapi (>=0.115.14,<0.116.0)",
    "black (>=25.1.0,<26.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)"
]

[tool.poetry]
packages = [{include = "src", from = "."}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
